abstracted decode and encode url for other purposes
